refactor(pane): split v1 frame impls into render/frames/v1/
Phase A of v1→v2 statusline redesign. Pure structural refactor with zero
behavior change — TOML strings (`none`/`zones`/`grid`/`cards`/`sections`)
and rendered output are unchanged.

- PaneStyle variants gain `V1` namespace prefix to leave room for v2 layout
  styles (cockpit/console/flightstrip/auto) shipping next release
- Per-frame render impls move from a 470-line pane.rs into one file each
  under src/render/frames/v1/ (zones/grid/cards/sections), with shared.rs
  for helpers used by 2+ frames
- pane.rs is now a slim ~120-line dispatcher owning the user-visible types
- New docs/pane-styles.md consolidates v1 frame reference with mockups
- TOML config templates mark v1 vs v2 namespace in their comments

Verified: cargo check / fmt / clippy clean, 300/300 tests pass, end-to-end
binary against tests/fixtures/minimal_statusline_input.json renders the
sections frame byte-identically to pre-refactor output.
