+++ src/render/mod.rs
@@ -1,5 +1,6 @@
 pub mod color;
 pub mod fmt;
+pub mod frames;
 pub mod icons;
 pub mod layout;
 pub mod pane;
